Улучшает структуру компонентов, начинает составлять структуру данных для точек маршрута
