+++ src/components/events-list/route-point/route-point-data.js
@@ -1,0 +1,39 @@
+const routeType = [`Taxi`, `Bus`, `Train`, `Ship`, `Transport`, `Drive`, `Flight`, `Check`, `Sightseeing`, `Restaurant`];
+const routeDestination = [`Copenhagen`, `Longyearbyen`, `Tromsø`, `Reykjavík`, `Nuuk`, `Moscow`, `Tórshavn`];
+
+const routeOffers = [{
+  title: `Add luggage + € `,
+  price: 30
+},
+{
+  title: `Switch to comfort class + €`,
+  price: 100
+},
+{
+  title: `Add meal + €`,
+  price: 15
+},
+{
+  title: `Choose seats + €`,
+  price: 5
+},
+{
+  title: `Travel by train + €`,
+  price: 40
+}];
+
+const routeDescription = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit. `,
+  `Cras aliquet varius magna, non porta ligula feugiat eget. `,
+  `Fusce tristique felis at fermentum pharetra. `,
+  `Aliquam id orci ut lectus varius viverra. `,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. `,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. `,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. `,
+  `Sed sed nisi sed augue convallis suscipit in sed felis. `,
+  `Aliquam erat volutpat. `,
+  `Nunc fermentum tortor ac porta dapibus. `,
+  `In rutrum ac purus sit amet tempus. `
+];
+
+export {routeType, routeDestination, routeOffers, routeDescription};
