+++ src/components/header/trip-cost/trip-cost.js
@@ -1,0 +1,5 @@
+export const renderTripCost = () => {
+  return `<p class="trip-info__cost">
+            Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+          </p>`;
+};
