+++ src/components/header/trip-info-container.js
