--- src/components/page-navigation.js
@@ -1,6 +1,0 @@
-export const renderPageNavigation = () => {
-  return `<nav class="trip-controls__trip-tabs  trip-tabs">
-  <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-  <a class="trip-tabs__btn" href="#">Stats</a>
-</nav>`;
-};
