+++ src/components/utils.js
@@ -1,0 +1,12 @@
+const getRandomIntegerNumber = (min, max) => {
+  return min + Math.floor(Math.random() * (max - min));
+};
+
+const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length);
+
+  return array[randomIndex];
+};
+
+
+export {getRandomIntegerNumber, getRandomArrayItem};
